requirements.txt
openai
numpy
matplotlib
python-dotenv